backend/requirements.txt
# FastAPI Backend Dependencies for SuperhireX
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database & Supabase
supabase==2.27.2

# AI & Document Processing
openai==1.99.9
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1

# CORS & Security
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.28.1
